// packages/vchart/src/interaction/triggers/interface.ts
export type ITriggerEventHandler = (e: any, markGraphic?: IMarkGraphic) => void;

export interface ITrigger<TriggerOptions extends IBaseTriggerOptions = any> {
  readonly options: TriggerOptions;
  readonly type: string;
  registerMark: (mark: IMark | IMark[]) => void;
  release: () => void;

  init: () => void;
  start: (g: any) => void;
  reset: (g?: IMarkGraphic) => void;
  getStartState: () => string;
  getResetState: () => string;
  updateMarkIdByState: (states: string[]) => void;
  getMarkIdByState: () => Record<string, number[]>;
  getMarks: () => IMark[];
  getMarksByState: (state: string) => IMark[];
}

export interface IBaseTriggerOptions {
  type?: string;
  /**
   * 需要处理状态的所有图元
   */
  marks?: IMark[];
  mode?: RenderMode;
  event: {
    on: (eType: string, callback: ITriggerEventHandler) => void;
    off: (eType: string, callback?: ITriggerEventHandler) => void;
    emit: (eType: string, params: any) => void;
  };
  interaction: IInteraction;

  id?: string;

  shouldStart?: (e: any) => boolean;

  shouldUpdate?: (e: any) => boolean;

  shouldEnd?: (e: any) => boolean;

  shouldReset?: (e: any) => boolean;

  onStart?: (e: any) => boolean;

  onUpdate?: (e: any) => boolean;

  onEnd?: (e: any) => boolean;

  onReset?: (e: any) => boolean;
}
// ...
export interface IDimensionHoverOptions extends IBaseTriggerOptions {
  /**
   * the selected state name
   */
  state?: string;
  /**
   * the non-selected state name
   */
  reverseState?: string;

  trigger?: string;
}